requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
pymysql
pandas
numpy
statsmodels
scikit-learn
Pillow
gunicorn
requests
matplotlib
